+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Versions follow [Semantic Versioning](https://semver.org/>) (<major>.<minor>.<patch>).
 
+### Added
+
+- Display Canaille debug logs by default.
+
 ## [0.0.6] - 2024-01-24
 
 ### Fixed